plugins/vcenter.py
import ssl
import logging
from typing import Dict, Any
from pyVim.connect import SmartConnect, Disconnect
from pyVmomi import vim
from plugins.base import PluginBase

logger = logging.getLogger(__name__)


class VCenterPlugin(PluginBase):
    """Plugin for VMware vCenter/ESXi management."""
    
    @property
    def plugin_type(self) -> str:
        return "vcenter"
    
    def test_connection(self, config: Dict[str, Any]) -> Dict[str, str]:
        host = config.get("host", "")
        username = config.get("api_key", "")
        password = config.get("api_endpoint", "")
        
        if not all([host, username, password]):
            missing = [k for k, v in {"host": host, "api_key": username, "api_endpoint": password}.items() if not v]
            logger.error(f"vCenter test_connection missing fields: {', '.join(missing)}")
            return {"status": "error", "details": "Missing required configuration"}
        
        try:
            logger.debug(f"Attempting vCenter connect: host={host}, user_present={bool(username)}, pwd_present={bool(password)}")
            try:
                pwd_len = len(password)
                pwd_trailing_ws = password.endswith((" ", "\t", "\r", "\n"))
                logger.debug(f"vCenter creds summary: user_len={len(username)}, pwd_len={pwd_len}, pwd_trailing_ws={pwd_trailing_ws}")
            except Exception:
                pass
            context = ssl._create_unverified_context()
            si = SmartConnect(
                host=host,
                user=username,
                pwd=password,
                port=443,
                sslContext=context
            )
            logger.info(f"vCenter connection established to {host}")
            Disconnect(si)
            return {"status": "online", "details": "Connection successful"}
        except Exception as e:
            msg = str(e)
            logger.error(f"vCenter test_connection error for {host}: {type(e).__name__}: {msg}")
            if "incorrect" in msg.lower() or "auth" in msg.lower() or "login" in msg.lower():
                return {"status": "auth_failed", "details": "Invalid credentials"}
            return {"status": "error", "details": msg}
    
    def execute_shutdown(self, config: Dict[str, Any]) -> Dict[str, Any]:
        host = config.get("host", "")
        username = config.get("api_key", "")
        password = config.get("api_endpoint", "")
        
        result = {"host": host, "status": "unknown", "details": ""}
        
        if not all([host, username, password]):
            missing = [k for k, v in {"host": host, "api_key": username, "api_endpoint": password}.items() if not v]
            logger.error(f"vCenter execute_shutdown missing fields: {', '.join(missing)}")
            result["status"] = "error"
            result["details"] = "Missing required configuration"
            return result
        
        try:
            logger.debug(f"Attempting vCenter shutdown: host={host}, user_present={bool(username)}, pwd_present={bool(password)}")
            try:
                pwd_len = len(password)
                pwd_trailing_ws = password.endswith((" ", "\t", "\r", "\n"))
                logger.debug(f"vCenter creds summary: user_len={len(username)}, pwd_len={pwd_len}, pwd_trailing_ws={pwd_trailing_ws}")
            except Exception:
                pass
            context = ssl._create_unverified_context()
            si = SmartConnect(
                host=host,
                user=username,
                pwd=password,
                port=443,
                sslContext=context
            )
            logger.info(f"vCenter connection established to {host} for shutdown phase")
            
            content = si.RetrieveContent()
            obj_view = content.viewManager.CreateContainerView(
                content.rootFolder,
                [vim.HostSystem],
                True
            )
            
            host_list = obj_view.view
            obj_view.Destroy()
            
            for host_obj in host_list:
                logger.info(f"Processing ESXi host: {host_obj.name}")
            
            result["status"] = "pending_implementation"
            result["details"] = "pyVmomi API requires documentation review"
            
            Disconnect(si)
            
        except Exception as e:
            msg = str(e)
            logger.error(f"vCenter execute_shutdown error for {host}: {type(e).__name__}: {msg}")
            result["status"] = "error"
            result["details"] = msg
        
        return result
    
    def get_required_fields(self) -> Dict[str, str]:
        return {
            "host": "vCenter hostname or IP",
            "api_key": "vSphere username",
            "api_endpoint": "vSphere password"
        }
