requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyotp==2.9.0
pyvmomi==8.0.2.0.1
python-jose==3.3.0
requests==2.31.0
cryptography==41.0.7
qrcode==7.4.2
colorama==0.4.6
paramiko==3.4.0
python-telegram-bot==20.7
python-dotenv==1.0.0
